+++ lib/src/Connection.dart
@@ -17,10 +17,13 @@ import 'package:xmpp_stone/src/presence/PresenceManager.dart';
 import 'package:xmpp_stone/src/roster/RosterManager.dart';
 import 'package:xmpp_stone/xmpp_stone.dart';
 
+import 'logger/Log.dart';
+
 enum XmppConnectionState {
   Idle,
   Closed,
   SocketOpening,
+  SocketOpened,
   DoneParsingFeatures,
   StartTlsFailed,
   AuthenticationNotSupported,
@@ -30,16 +33,20 @@ enum XmppConnectionState {
   AuthenticationFailure,
   Resumed,
   SessionInitialized,
-  DoneServiceDiscovery,
   Ready,
-  ForcelyClosed,
-  Reconnecting
+  Closing,
+  ForcefullyClosed,
+  Reconnecting,
+  WouldLikeToOpen,
+  WouldLikeToClose,
 }
 
 class Connection {
   var lock = Lock(reentrant: true);
 
-  static Map<String, Connection> instances = Map<String, Connection>();
+  static String TAG = 'Connection';
+
+  static Map<String, Connection> instances = <String, Connection>{};
 
   XmppAccountSettings account;
 
@@ -55,8 +62,8 @@ class Connection {
 
   String _serverName;
 
-  static getInstance(XmppAccountSettings account) {
-    Connection connection = instances[account.fullJid.userAtDomain];
+  static Connection getInstance(XmppAccountSettings account) {
+    var connection = instances[account.fullJid.userAtDomain];
     if (connection == null) {
       connection = Connection(account);
       instances[account.fullJid.userAtDomain] = connection;
@@ -74,19 +81,19 @@ class Connection {
 
   bool authenticated = false;
 
-  StreamController<AbstractStanza> _inStanzaStreamController =
+  final StreamController<AbstractStanza> _inStanzaStreamController =
       StreamController.broadcast();
 
-  StreamController<AbstractStanza> _outStanzaStreamController =
+  final StreamController<AbstractStanza> _outStanzaStreamController =
       StreamController.broadcast();
 
-  StreamController<Nonza> _inNonzaStreamController =
+  final StreamController<Nonza> _inNonzaStreamController =
       StreamController.broadcast();
 
-  StreamController<Nonza> _outNonzaStreamController =
+  final StreamController<Nonza> _outNonzaStreamController =
       StreamController.broadcast();
 
-  StreamController<XmppConnectionState> _connectionStateStreamController =
+  final StreamController<XmppConnectionState> _connectionStateStreamController =
       StreamController.broadcast();
 
   Stream<AbstractStanza> get inStanzasStream {
@@ -109,11 +116,9 @@ class Connection {
     return _connectionStateStreamController.stream;
   }
 
-  bool _logXML = true;
-
   Jid get fullJid => account.fullJid;
 
-  ConnectionNegotatiorManager connectionNegotatiorManager;
+  ConnectionNegotiatorManager connectionNegotatiorManager;
 
   void fullJidRetrieved(Jid jid) {
     account.resource = jid.resource;
@@ -135,59 +140,57 @@ class Connection {
     PresenceManager.getInstance(this);
     MessageHandler.getInstance(this);
     PingManager.getInstance(this);
+    connectionNegotatiorManager = ConnectionNegotiatorManager(this, account);
     reconnectionManager = ReconnectionManager(this);
     _logXML = enableLog;
   }
 
   void _openStream() {
-    String streamOpeningString = """
+    var streamOpeningString = """
 <?xml version='1.0'?>
 <stream:stream xmlns='jabber:client' version='1.0' xmlns:stream='http://etherx.jabber.org/streams'
 to='${fullJid.domain}'
 xml:lang='en'
 >
 """;
-
     write(streamOpeningString);
   }
 
-  String restOfResponse = "";
+  String restOfResponse = '';
 
   String extractWholeChild(String response) {
     return response;
   }
 
   String prepareStreamResponse(String response) {
-    String response1 = extractWholeChild(restOfResponse + response);
-
-    if (response1.contains("stream:stream") &&
-        !(response1.contains("</stream>"))) {
+    Log.xmppp_receiving(response);
+    var response1 = extractWholeChild(restOfResponse + response);
+    if (response1.contains('</stream:stream>')) {
+      close();
+      return '';
+    }
+    if (response1.contains('stream:stream') &&
+        !(response1.contains('</stream:stream>'))) {
       response1 = response1 +
-          "</stream>"; // fix for crashing xml library without ending
+          '</stream:stream>'; // fix for crashing xml library without ending
     }
 
     //fix for multiple roots issue
-    response1 = "<xmpp_stone>$response1</xmpp_stone>";
-
-    if (_logXML) {
-      print("response: ${response1}");
-    }
-    if (response1.contains("</stream:stream>")) {
-      close();
-      return "";
-    }
+    response1 = '<xmpp_stone>$response1</xmpp_stone>';
     return response1;
   }
 
   void reconnect() {
-    print("Reconnect!!!");
-    if (_state == XmppConnectionState.ForcelyClosed) {
+    if (_state == XmppConnectionState.ForcefullyClosed) {
       setState(XmppConnectionState.Reconnecting);
       openSocket();
     }
   }
 
   void connect() {
+    if (_state == XmppConnectionState.Closing) {
+      _state == XmppConnectionState.WouldLikeToOpen;
+    }
     if (_state == XmppConnectionState.Closed) {
       _state = XmppConnectionState.Idle;
     }
@@ -197,97 +200,106 @@ xml:lang='en'
   }
 
   Future<void> openSocket() async {
-    connectionNegotatiorManager =
-        ConnectionNegotatiorManager(this, account.password);
+    connectionNegotatiorManager.init();
+    setState(XmppConnectionState.SocketOpening);
     try {
-      return await Socket.connect(account.host ?? account.domain, account.port).then((Socket socket) {
-        _socket = socket;
-        socket
-            .cast<List<int>>()
-            .transform(utf8.decoder)
-            .map(prepareStreamResponse)
-            .listen(handleResponse,
-            onDone: handleConnectionDone);
-        _openStream();
-        setState(XmppConnectionState.SocketOpening);
+      return await Socket.connect(account.host ?? account.domain, account.port)
+          .then((Socket socket) {
+        // if not closed in meantime
+        if (_state != XmppConnectionState.Closed) {
+          setState(XmppConnectionState.SocketOpened);
+          _socket = socket;
+          socket
+              .cast<List<int>>()
+              .transform(utf8.decoder)
+              .map(prepareStreamResponse)
+              .listen(handleResponse, onDone: handleConnectionDone);
+          _openStream();
+        } else {
+          Log.d(TAG, 'Closed in meantime');
+          socket.close();
+        }
       });
-    } on SocketException catch(error) {
-      print("Socket Exception" + error.toString());
+    } on SocketException catch (error) {
+      Log.e(TAG, 'Socket Exception' + error.toString());
       handleConnectionError(error.toString());
     }
   }
 
   void close() {
-    if (state != XmppConnectionState.Closed) {
-      if (state != XmppConnectionState.ForcelyClosed) {
-        if (_socket != null) {
-          try {
-            _socket.write('</stream:stream>');
-            _socket.close();
-          } on Exception {
-            print("Socket already closed");
-          }
+    if (state == XmppConnectionState.SocketOpening) {
+      throw Exception('Closing is not possible during this state');
+    }
+    if (state != XmppConnectionState.Closed &&
+        state != XmppConnectionState.ForcefullyClosed &&
+        state != XmppConnectionState.Closing) {
+      if (_socket != null) {
+        try {
+          setState(XmppConnectionState.Closing);
+          _socket.write('</stream:stream>');
+        } on Exception {
+          Log.d(TAG, 'Socket already closed');
         }
       }
-      setState(XmppConnectionState.Closed);
       authenticated = false;
     }
   }
 
   bool startMatcher(xml.XmlElement element) {
-    String name = element.name.local;
-    return name == "stream";
+    var name = element.name.local;
+    return name == 'stream';
   }
 
   bool stanzaMatcher(xml.XmlElement element) {
-    String name = element.name.local;
-    return name == "iq" || name == "message" || name == "presence";
+    var name = element.name.local;
+    return name == 'iq' || name == 'message' || name == 'presence';
   }
 
   bool nonzaMatcher(xml.XmlElement element) {
-    String name = element.name.local;
-    return name != "iq" && name != "message" && name != "presence";
+    var name = element.name.local;
+    return name != 'iq' && name != 'message' && name != 'presence';
   }
 
   bool featureMatcher(xml.XmlElement element) {
-    String name = element.name.local;
-    return (name == "stream:features" || name == "features");
+    var name = element.name.local;
+    return (name == 'stream:features' || name == 'features');
   }
 
-  String _unparsedXmlResponse = "";
+  String _unparsedXmlResponse = '';
 
   void handleResponse(String response) {
     String fullResponse;
     if (_unparsedXmlResponse.isNotEmpty) {
-      print(_unparsedXmlResponse);
-      fullResponse = "$_unparsedXmlResponse${response.substring(12)}"; //
-      // remove xmpp_stone start tag
-      print(_unparsedXmlResponse);
-      _unparsedXmlResponse = "";
+      if (response.length > 12) {
+        fullResponse = '$_unparsedXmlResponse${response.substring(12)}'; //
+      } else {
+        fullResponse = _unparsedXmlResponse;
+      }
+      Log.v(TAG, 'full response = ${fullResponse}');
+      _unparsedXmlResponse = '';
     } else {
       fullResponse = response;
     }
 
     if (fullResponse != null && fullResponse.isNotEmpty) {
-      var xmlResponse;
+      xml.XmlNode xmlResponse;
       try {
-        //print(fullResponse);
-        xmlResponse = xml.parse(fullResponse).firstChild;
+        xmlResponse = xml.XmlDocument.parse(fullResponse).firstChild;
       } catch (e) {
         _unparsedXmlResponse += fullResponse.substring(
             0, fullResponse.length - 13); //remove  xmpp_stone end tag
-        xmlResponse = xml.XmlElement(xml.XmlName("error"));
+        xmlResponse = xml.XmlElement(xml.XmlName('error'));
       }
 //      xmlResponse.descendants.whereType<xml.XmlElement>().forEach((element) {
-//        print("element: " + element.name.local);
+//        Log.d("element: " + element.name.local);
 //      });
-
+      //TODO: Improve parser for children only
       xmlResponse.descendants
           .whereType<xml.XmlElement>()
           .where((element) => startMatcher(element))
           .forEach((element) => processInitialStream(element));
 
-      xmlResponse.descendants
+      xmlResponse.children
           .whereType<xml.XmlElement>()
           .where((element) => stanzaMatcher(element))
           .map((xmlElement) => StanzaParser.parseStanza(xmlElement))
@@ -299,21 +311,18 @@ xml:lang='en'
           .forEach((feature) =>
               connectionNegotatiorManager.negotiateFeatureList(feature));
 
-      xmlResponse.descendants
+      //TODO: Probably will introduce bugs!!!
+      xmlResponse.children
           .whereType<xml.XmlElement>()
           .where((element) => nonzaMatcher(element))
           .map((xmlElement) => Nonza.parse(xmlElement))
           .forEach((nonza) => _inNonzaStreamController.add(nonza));
-
-//      if (xmlResponse.findAllElements("stream:error").isNotEmpty) {
-//        processError(xmlResponse);
-//      }
     }
   }
 
   void processInitialStream(xml.XmlElement initialStream) {
-    print("processInitialStream");
-    String from = initialStream.getAttribute('from');
+    Log.d(TAG, 'processInitialStream');
+    var from = initialStream.getAttribute('from');
     if (from != null) {
       _serverName = from;
     }
@@ -321,13 +330,13 @@ xml:lang='en'
 
   bool isOpened() {
     return state != XmppConnectionState.Closed &&
-        state != XmppConnectionState.ForcelyClosed;
+        state != XmppConnectionState.ForcefullyClosed &&
+        state != XmppConnectionState.Closing &&
+        state != XmppConnectionState.SocketOpening;
   }
 
   void write(message) {
-    if (_logXML) {
-      print("sending: " + message);
-    }
+    Log.xmppp_sending(message);
     if (isOpened()) {
       _socket.write(message);
     }
@@ -347,7 +356,7 @@ xml:lang='en'
     _state = state;
     _fireConnectionStateChangedEvent(state);
     _processState(state);
-    print("State: ${_state}");
+    Log.d(TAG, 'State: ${_state}');
   }
 
   XmppConnectionState get state {
@@ -366,8 +375,7 @@ xml:lang='en'
   }
 
   void startSecureSocket() {
-    print("startSecureSocket");
-    print(state);
+    Log.d(TAG, 'startSecureSocket');
     SecureSocket.secure(_socket, onBadCertificate: _validateBadCertificate)
         .then((secureSocket) {
       _socket = secureSocket;
@@ -376,7 +384,9 @@ xml:lang='en'
           .transform(utf8.decoder)
           .map(prepareStreamResponse)
           .listen(handleResponse,
-              onError: (error) => {handleSecuredConnectionError(error.toString())}, onDone: handleSecuredConnectionDone);
+              onError: (error) =>
+                  {handleSecuredConnectionError(error.toString())},
+              onDone: handleSecuredConnectionDone);
       _openStream();
     });
   }
@@ -404,8 +414,9 @@ xml:lang='en'
   }
 
   void doneParsingFeatures() {
-    print("DONE PARSING FATURES");
-    setState(XmppConnectionState.DoneParsingFeatures);
+    if (state == XmppConnectionState.SessionInitialized) {
+      setState(XmppConnectionState.Ready);
+    }
   }
 
   void startTlsFailed() {
@@ -422,30 +433,37 @@ xml:lang='en'
   }
 
   void handleConnectionDone() {
-    print("!!!!!!!!!!!Handle connection done");
-    if (state != XmppConnectionState.Closed) {
-      setState(XmppConnectionState.ForcelyClosed);
-    }
+    Log.d(TAG, 'Handle connection done');
+    handleCloseState();
   }
 
   void handleSecuredConnectionDone() {
-    print("!!!!!!!!!!!Handle secured connection done");
-    if (state != XmppConnectionState.Closed) {
-      setState(XmppConnectionState.ForcelyClosed);
+    Log.d(TAG, 'Handle secured connection done');
+    handleCloseState();
+  }
+
+  void handleConnectionError(String error) {
+    handleCloseState();
+  }
+
+  void handleCloseState() {
+    if (state == XmppConnectionState.WouldLikeToOpen) {
+      setState(XmppConnectionState.Closed);
+      connect();
+    } else if (state != XmppConnectionState.Closing) {
+      setState(XmppConnectionState.ForcefullyClosed);
+    } else {
+      setState(XmppConnectionState.Closed);
     }
   }
 
-  handleConnectionError(String error) {
-    print("!!!!!!Handle ERROR " + error);
-    if (state != XmppConnectionState.Closed) {
-      setState(XmppConnectionState.ForcelyClosed);
-    }
+  void handleSecuredConnectionError(String error) {
+    Log.d(TAG, 'Handle Secured Error  $error');
+    handleCloseState();
   }
 
-  handleSecuredConnectionError(String error) {
-    print("!!!!!!Handle Secured ERROR " + error);
-    if (state != XmppConnectionState.Closed) {
-      setState(XmppConnectionState.ForcelyClosed);
-    }
+  bool isAsyncSocketState() {
+    return state == XmppConnectionState.SocketOpening ||
+        state == XmppConnectionState.Closing;
   }
 }