+++ lib/src/ReconnectionManager.dart
@@ -1,9 +1,10 @@
-
 import 'dart:async';
-
 import 'package:xmpp_stone/src/Connection.dart';
+import 'logger/Log.dart';
 
 class ReconnectionManager {
+  static const TAG = 'ReconnectionManager';
+
   Connection _connection;
   bool isActive = false;
   int initialTimeout = 1000;
@@ -17,15 +18,16 @@ class ReconnectionManager {
     _connection.connectionStateStream.listen(connectionStateHandler);
     initialTimeout = _connection.account.reconnectionTimeout;
     totalReconnections = _connection.account.totalReconnections;
+    timeOutInMs = initialTimeout;
   }
 
   void connectionStateHandler(XmppConnectionState state) {
-    if (state == XmppConnectionState.ForcelyClosed) {
-      print("Connection forcely closed!"); //connection lost
+    if (state == XmppConnectionState.ForcefullyClosed) {
+      Log.d(TAG, 'Connection forcefully closed!'); //connection lost
       handleReconnection();
     } else if (state == XmppConnectionState.SocketOpening) {
       //do nothing
-    }else if (state != XmppConnectionState.Reconnecting) {
+    } else if (state != XmppConnectionState.Reconnecting) {
       isActive = false;
       timeOutInMs = initialTimeout;
       counter = 0;
@@ -40,10 +42,10 @@ class ReconnectionManager {
     if (timer != null) {
       timer.cancel();
     }
-    if (counter< totalReconnections) {
+    if (counter < totalReconnections) {
       timer = Timer(Duration(milliseconds: timeOutInMs), _connection.reconnect);
       timeOutInMs += timeOutInMs;
-      print("TimeOut is:" + timeOutInMs.toString() + "recconection counter" + counter.toString());
+      Log.d(TAG, 'TimeOut is: $timeOutInMs reconnection counter $counter');
       counter++;
     } else {
       _connection.close();