+++ lib/src/elements/stanzas/PresenceStanza.dart
@@ -18,7 +18,7 @@ class PresenceStanza extends AbstractStanza {
   }
 
   PresenceType get type {
-    String typeValue = getAttribute('type')?.value;
+    var typeValue = getAttribute('type')?.value;
     return typeFromString(typeValue);
   }
 
@@ -35,7 +35,7 @@ class PresenceStanza extends AbstractStanza {
   //status with no language prefs
   String get status {
     var statusElement =
-        children.firstWhere((element) => element.name == name && element.attributes.isEmpty, orElse: () => null);
+        children.firstWhere((element) => element.name == 'status' && element.attributes.isEmpty, orElse: () => null);
     return statusElement?.textValue;
   }