+++ lib/src/features/Negotiator.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:xmpp_stone/src/elements/nonzas/Nonza.dart';
 
-abstract class ConnectionNegotiator {
+abstract class Negotiator {
   static int defaultPriorityLevel = 1000;
 
   String _expectedName;
@@ -13,7 +13,7 @@ abstract class ConnectionNegotiator {
   NegotiatorState get state => _state;
 
   StreamController<NegotiatorState> negotiatorStateStreamController =
-      StreamController();
+      StreamController<NegotiatorState>.broadcast();
 
   Stream<NegotiatorState> get featureStateStream {
     return negotiatorStateStreamController.stream;
@@ -39,11 +39,15 @@ abstract class ConnectionNegotiator {
     _expectedName = value;
   }
 
-  bool match(Nonza request);
+  //goes trough all features and match only needed nonzas
+  List<Nonza> match(List<Nonza> request);
 
-  void negotiate(Nonza nonza);
+  void negotiate(List<Nonza> nonza);
 
   void backToIdle() {
+    if (negotiatorStateStreamController.isClosed) {
+      negotiatorStateStreamController = StreamController();
+    }
     state = NegotiatorState.IDLE;
   }